dts: fix NOR flash
Ewald provided a working example of how the flash is supposed to work.
Turned out the "jedec-flash" is the correct binding. With it, it works!

He also reported the following error:
"platform physmap-flash.0: failed to claim resource 0: [mem 0x08000000-0x07ffffff]"
This was caused by CONFIG_MTD_PHYSMAP_COMPAT, because
it expects a compat phymap to be setup through the
kernel configuration. (which has default bogus data).

Suggested-by: Ewald Comhaire <e.comhaire@gmail.com>
Signed-off-by: Christian Lamparter <chunkeey@gmail.com>