build: add support to boot from USB for the MyBook Live DUO.
Details are in the README.md. Instructions are written for the
UART-Cable. While it's certainly possible to do this without
attaching the UART, it's so much more dangerous... since in case
you brick it, you can easily restore it with the cable already
attached, than without...

No, the MyBook Live (Single) doesn't have an USB-Port, so skip
this if you only the the "Single".

Signed-off-by: Christian Lamparter <chunkeey@gmail.com>