root: rename and rework disable-ro/enable-ro/rw scripts
the previous code was confusing me as well.

So, the mbl-debian starts out with root "/" in "read-write"-mode as usual.
The user then can use enable-ro.sh (which reboots the NAS too), so that on
the next boot the root is read-only and an overlayfs is mounted that keeps
all the filesystem changes in a tmpfs.

Signed-off-by: Christian Lamparter <chunkeey@gmail.com>