debug: add debug-image utility
With ports being ports, there's a lot of churn how the current sid/unstable
packages can't be build for powerpc. This utility allows one to >enter< a
generated image and either confirm what's broken or how to come up with an
workaround.

Due to the commands this utility calls, it needs to be run as root or some
account that has access to kpartx + mount + chroot.

Signed-off-by: Christian Lamparter <chunkeey@gmail.com>